Labelling issue
Fixed tier labelling issue where in the sliced elan it would not cut
annotations but some would extend out of the picture to the left, and
the tier label position was tied to that and hence outside the visible
area. Now label position is tied to elan.range(1) and thus always next
to the axis.