Bug fix to elanRenameTiers
1) bug caused the structure to change (output did not have .tiers. level
2) now you can use the function as a tier name unifier by submitting
full lists of tier names. It renames the ones that are different and
keeps the rest. In the old function if you had the same tier name in
new as in the old list, it would be deleted. This change can be useful
if you want to unify spelling of tier names across different files,
just extract the current list from the file using fieldnames.m and