Two new functions
elanExtend lengthens the annotations by a given time value. This is for
example useful when needing to search for stuff that happens right
after or around events at a tier. Extend the original annotations and
then use the elanCorrelate function to find overlaps. Previously
non-functioning elanRemoveLabel now drops the annotations that have one
or more of the listed values. This is handy in cleaning up the data, or
selecting only a subset of events for plotting etc.