Update to plotting and tier correlations
This update contains an update to plotting tiers (making colour-codes
for each annotation value using elanAssociate and then plotting the
tiers with elanPlot). Also contains new function elanCorrCheck that
counts the number of simultaneous occurrences of annotations in two
tiers. The tier correlations function elanCorrelateTiers (and its
helper elanComputeOverlap) for generating new tiers based on overlaps
of annotations are copied as-is from the SALEM toolbox.